✨ Encrypt Ballot Endpoint (#31)
* ✨ Encrypt Ballot Endpoint

Encrypt a single ballot via endpoint. 
Provide new ways to pass in the ElectionMap and ElectionGuardConnfig to remove the restrictions from coming into the pathway from different directions. This also allows for a different direction for the current count which now can be passed in OR just simply incremented.

* Rework the mock EncryptBallot API

add support for loading configuration at runtime from a relative path

* Feature/modify controller endpoint (#34)

* Change route from /election to /electionguard

* check the call context for data directory

if failed check the parent folder

Co-authored-by: Matt Wilhelm <AddressXception@users.noreply.github.com>